requirements_k2450.txt
# Requirements for K2450 Control Suite
# Install with: pip install -r requirements_k2450.txt

# GUI Framework
PyQt5>=5.15.0
pyqtgraph>=0.13.0

# Instrument Communication
pyvisa>=1.13.0
pyvisa-py>=0.7.0         # Pure Python VISA backend (for USB without NI-VISA)
pyusb>=1.2.0             # USB communication support

# Scientific Computing
numpy>=1.24.0

# Image processing (for icon generation during development)
Pillow>=10.0.0

# Build tools (for creating distributable)
pyinstaller>=6.0.0
